Started refactoring
