+++ __tests__/Logger.test.ts
@@ -1,9 +1,6 @@
 import {Logger} from '../src/share/Logger'
-import {Reactor, InPort, Read, Triggers, Args, State, Present, ReactionSandbox, App} from '../src/core/reactor';
-import { TimeValue, TimeUnit } from '../src/core/time';
+import {Reactor, App} from '../src/core/reactor';
 import { Log, LogLevel } from '../src/core/util'
-
-
 
 const _reactor:Reactor = new App()
 const lg:Logger = new Logger(_reactor , 10)
