+++ src/core/container.ts
@@ -1,0 +1,11 @@
+import {Component} from './component';
+import { Reaction } from './reaction';
+import { Port, Present } from './reactor';
+import { DependencyGraph } from './util';
+
+export abstract class Container extends Component {
+
+    
+
+
+}
