+++ src/core/reactor.ts
@@ -10,6 +10,10 @@ import {PrioritySetElement, PrioritySet, SortableDependencyGraph, Log, Dependenc
 import {TimeValue, TimeUnit, Tag, Origin, getCurrentPhysicalTime, Alarm} from './time';
 import {Component} from "./component"
 import {Reaction, Priority, Mutation, Procedure} from "./reaction"
+import { IOPort, MultiPort, Port, WritablePort } from './port';
+import { Action, Shutdown, Startup } from './action';
+import { ScheduledTrigger, Trigger, TriggerManager } from './trigger';
+import { TaggedEvent } from './event';
 
 // Set the default log level.
 Log.global.level = Log.levels.ERROR;
@@ -53,11 +57,6 @@ export type ReadWrite<T> = Read<T> & Write<T>;
  */
 export type Variable = Read<unknown> | Array<Read<unknown>> | MultiPort<Present>
 
-//--------------------------------------------------------------------------//
-// Constants                                                                //
-//--------------------------------------------------------------------------//
-
-const defaultMIT = TimeValue.withUnits(1, TimeUnit.nsec); // FIXME
 
 //--------------------------------------------------------------------------//
 // Interfaces                                                               //
@@ -92,16 +91,7 @@ export interface Write<T> {
     set: (value: T) => void;
 }
 
-/**
- * Abstract class for a writable port. It is intended as a wrapper for a
- * regular port. In addition to a get method, it also has a set method and
- * a method for retrieving the port that it wraps.
- */
-export abstract class WritablePort<T extends Present> implements ReadWrite<T> {
-    abstract get(): T | undefined;
-    abstract set(value: T): void;
-    abstract getPort(): IOPort<T>
-}
+
 
 /**
  * Abstract class for a schedulable action. It is intended as a wrapper for a
@@ -117,309 +107,6 @@ export abstract class SchedulableAction<T extends Present> implements Sched<T> {
 // Core Reactor Classes                                                     //
 //--------------------------------------------------------------------------//
 
-/**
- * An event is caused by a timer or a scheduled action. Each event is tagged
- * with a time instant and may carry a value of arbitrary type. The tag will
- * determine the event's position with respect to other events in the event
- * queue.
- */
-export class TaggedEvent<T extends Present> implements PrioritySetElement<Tag> {
-
-    /**
-     * Pointer to the next element of the priority set that this event might
-     * be hooked into.
-     */
-    public next: PrioritySetElement<Tag> | undefined;
-
-    /**
-     * Construct a new tagged event.
-     * @param trigger The trigger of this event.
-     * @param tag The tag at which this event occurs.
-     * @param value The value associated with this event. 
-     * 
-     */
-    constructor(public trigger: ScheduledTrigger<T>, public tag: Tag, public value: T) {
-    }
-
-    /**
-     * Return true if this event has a smaller tag than the given event, false
-     * otherwise.
-     * @param node The event to compare this event's tag against.
-     */
-    hasPriorityOver(node: PrioritySetElement<Tag> | undefined) {
-        if (node) {
-            return this.getPriority().isSmallerThan(node.getPriority());
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * Determine whether the given event is a duplicate of this one. If so, assign the
-     * value this event to the given one. Otherwise, return false.
-     * @param node The event adopt the value from if it is a duplicate of this one.
-     */
-    updateIfDuplicateOf(node: PrioritySetElement<Tag> | undefined) {
-        if (node && node instanceof TaggedEvent) {
-            if (this.trigger === node.trigger && this.tag.isSimultaneousWith(node.tag)) {
-                node.value = this.value; // update the value
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Return the tag associated with this event.
-     */
-    getPriority(): Tag {
-        return this.tag;
-    }
-}
-
-/**
- * Abstract class for a trigger. A trigger may be an action, port, or timer.
- */
-abstract class Trigger extends Component {
-
-    /**
-     * Reactions to trigger.
-     */
-    protected reactions: Set<Reaction<unknown>> = new Set();
-    
-    /**
-     * Request the manager of this trigger. The request will only be honored
-     * if the correct key is given. Each component has a unique symbol (a key)
-     * that is handed to the owner upon instantiation of the component. If the
-     * wrong key is supplied, return undefined.
-     * @param key The private key embedded in this trigger.
-     */
-    abstract getManager(key: Symbol | undefined): TriggerManager;
-
-    /**
-     * Return the owner of this trigger.
-     */
-    public getContainer(): Reactor | null {
-        return this._getContainer()
-    }
-
-    /**
-     * Return whether or not this trigger is present.
-     */
-    abstract isPresent(): boolean;
-
-}
-
-/**
- * 
- */
-abstract class ScheduledTrigger<T extends Present> extends Trigger {
-    protected value: T | Absent = undefined;
-    protected tag: Tag | undefined;
-    
-    protected runtime!: Runtime;
-
-    constructor(container: Reactor, alias?: string) { // FIXME: do we really want the alias here? Probably not.
-        super(container, alias)
-        this._linkToRuntimeObject()
-    }
-
-    /**
-     * Update the current value of this timer in accordance with the given
-     * event, and trigger any reactions that list this timer as their trigger.
-     * @param e Timestamped event.
-     */
-    public update(e: TaggedEvent<T>):void {
-
-        if (!e.tag.isSimultaneousWith(this.runtime.util.getCurrentTag())) {
-            throw new Error("Time of event does not match current logical time.");
-        }
-        if (e.trigger === this) {
-            this.value = e.value
-            this.tag = e.tag;
-            for (let r of this.reactions) {
-                this.runtime.stage(r)
-            }
-        } else {
-            throw new Error("Attempt to update action using incompatible event.");
-        }
-    }
-
-    public getManager(key: Symbol | undefined): TriggerManager {
-        if (this._key == key) {
-            return this.manager
-        }
-        throw Error("Unable to grant access to manager.")
-    }
-
-    /**
-     * Returns true if this action was scheduled for the current
-     * logical time. This result is not affected by whether it
-     * has a value.
-     */
-    public isPresent() {
-        if (this.tag === undefined) {
-            // This action has never been scheduled before.
-            return false;
-        }
-        if (this.tag.isSimultaneousWith(this.runtime.util.getCurrentTag())) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    protected manager = new class implements TriggerManager {
-        constructor(private trigger: ScheduledTrigger<T>) { }
-        getContainer(): Reactor {
-            return this.trigger._getContainer()
-        }
-        addReaction(reaction: Reaction<unknown>): void {
-            this.trigger.reactions.add(reaction)
-        }
-        delReaction(reaction: Reaction<unknown>): void {
-            this.trigger.reactions.delete(reaction)
-        }
-    }(this)
-
-    public _receiveRuntimeObject(runtime: Runtime) {
-        if (!this.runtime) {
-            this.runtime = runtime
-        } else {
-            throw new Error("Can only establish link to runtime once.")
-        }
-    }
-
-}
-
-/**
- * An action denotes a self-scheduled event.
- * An action, like an input, can cause reactions to be invoked.
- * Whereas inputs are provided by other reactors, actions are scheduled
- * by this reactor itself, either in response to some observed external
- * event or as a delayed response to some input event. The action can be
- * scheduled by a reactor by invoking the schedule function in a reaction
- * or in an asynchronous callback that has been set up in a reaction.
- */
-export class Action<T extends Present> extends ScheduledTrigger<T> implements Read<T> {
-
-    readonly origin: Origin;
-    readonly minDelay: TimeValue;
-    readonly minInterArrival: TimeValue = defaultMIT;
-    
-    public get(): T | Absent {
-        if (this.isPresent()) {
-            return this.value;
-        } else {
-            return undefined;
-        }
-    }
-
-    public asSchedulable(key: Symbol | undefined): Sched<T> {
-        if (this._key === key) {
-            return this.scheduler
-        }
-        throw Error("Invalid reference to container.")
-    }
-
-    public getManager(key: Symbol | undefined): TriggerManager {
-        if (this._key == key) {
-            return this.manager
-        }
-        throw Error("Unable to grant access to manager.")
-    }
-
-    protected scheduler = new class<T extends Present> extends SchedulableAction<T> {
-        get(): T | undefined {
-            return this.action.get()
-        }
-        constructor(private action: Action<T>) {
-            super()
-        }
-        schedule(extraDelay: 0 | TimeValue, value: T, intendedTag?: Tag): void {
-            if (!(extraDelay instanceof TimeValue)) {
-                extraDelay = TimeValue.secs(0);
-            }
-            
-            var tag = this.action.runtime.util.getCurrentTag();
-            var delay = this.action.minDelay.add(extraDelay);
-
-            tag = tag.getLaterTag(delay);
-
-            if (this.action.origin == Origin.physical) {
-                // If the resulting timestamp from delay is less than the current physical time
-                // on the platform, then the timestamp becomes the current physical time.
-                // Otherwise the tag is computed like a logical action's tag.
-
-                let physicalTime = getCurrentPhysicalTime();
-                if (tag.time.isEarlierThan(physicalTime)) {
-                    tag = new Tag(getCurrentPhysicalTime(), 0);
-                } else {
-                    tag = tag.getMicroStepLater();
-                }
-            }
-
-            if (this.action instanceof FederatePortAction) {
-                if (intendedTag === undefined) {
-                    throw new Error("FederatedPortAction must have an intended tag from RTI.");
-                }
-                if (intendedTag <= this.action.runtime.util.getCurrentTag()) {
-                    throw new Error("Intended tag must be greater than current tag. Intended tag" +
-                    intendedTag + " Current tag: " + this.action.runtime.util.getCurrentTag());
-                }
-                Log.debug(this, () => "Using intended tag from RTI, similar to schedule_at_tag(tag) with an intended tag: " +
-                intendedTag);
-                tag = intendedTag;
-            } else if (this.action.origin == Origin.logical && !(this.action instanceof Startup)) {
-                tag = tag.getMicroStepLater();
-            }
-            
-            Log.debug(this, () => "Scheduling " + this.action.origin +
-                " action " + this.action._getFullyQualifiedName() + " with tag: " + tag);
-    
-            this.action.runtime.schedule(new TaggedEvent(this.action, tag, value));
-        }
-    }(this)
-
-    /** 
-     * Construct a new action.
-     * @param __container__ The reactor containing this action.
-     * @param origin Optional. If physical, then the hardware clock on the local 
-     * platform is used to determine the tag of the resulting event. If logical, 
-     * the current logical time (plus one microstep) is used as the offset.
-     * @param minDelay Optional. Defaults to 0. Specifies the intrinsic delay of
-     * any events resulting from scheduling this action.
-     * @param minInterArrival Optional. Defaults to 1 nsec. Specifies the minimum
-     * intrinsic delay between to occurrences of this action.
-     */
-    constructor(__container__: Reactor, origin: Origin, minDelay: TimeValue = TimeValue.secs(0), minInterArrival: TimeValue = defaultMIT) {
-        super(__container__);
-        this.origin = origin;
-        this.minDelay = minDelay;
-    }
-
-    public toString() {
-        return this._getFullyQualifiedName();
-    }
-}
-
-export class Startup extends Action<Present> { // FIXME: this should not be a schedulable trigger, just a trigger
-    constructor(__parent__: Reactor) {
-        super(__parent__, Origin.logical)
-    }
-}
-
-export class Shutdown extends Action<Present> {
-    constructor(__parent__: Reactor) {
-        super(__parent__, Origin.logical)
-    }
-}
-
-export class FederatePortAction extends Action<Buffer> {
-    constructor(__parent__: Reactor) {
-        super(__parent__, Origin.logical)
-    }
-}
 
 export class Parameter<T> implements Read<T> {
     constructor(private value: T) {
@@ -429,37 +116,7 @@ export class Parameter<T> implements Read<T> {
     }
 }
 
-/**
- * A state variable. This class refines the Read interface by letting `get`
- * return T rather than T | Absent. If the state should be nullable or
- * uninitialized, this has to be reflected explicitly in T.
- */
-export class State<T> implements Read<T>, Write<T> {
-    
-    /**
-     * Create a new state variable and assign it an initial value.
-     * @param value The initial value to assign to this state variable.
-     */
-    constructor(private value: T) {
-    
-    }
 
-    /**
-     * Return the current value of this state variable.
-     */
-    get(): T {
-        return this.value;
-    };
-
-    /**
-     * Set the current value of this state variable.
-     * @param value 
-     */
-    set(value: T) {
-        this.value = value;
-    };
-
-}
 
 /**
  * A timer is an attribute of a reactor which periodically (or just once)
@@ -1703,242 +1360,18 @@ protected _getFirstReactionOrMutation(): Reaction<any> | undefined {
     }
 }
 
-export abstract class Port<T extends Present> extends Trigger implements Read<T> {
-    
-    protected runtime!: Runtime;
 
-    constructor(container: Reactor, alias?: string) {
-        super(container, alias)
-        this._linkToRuntimeObject()
-    }
 
-    /** The time stamp associated with this port's value. */
-    protected tag: Tag | undefined;
 
-    /** The value associated with this port. */
-    protected value: T | Absent;
 
-    abstract get(): T | undefined;
-
-    public _receiveRuntimeObject(runtime: Runtime) {
-        if (!this.runtime) {
-            this.runtime = runtime
-        } else {
-            throw new Error("Can only establish link to runtime once. Name: " + this._getFullyQualifiedName())
-        }
-    }
-
-    /**
-     * Returns true if the connected port's value has been set; false otherwise
-     */
-    public isPresent() {
-
-        Log.debug(this, () => "In isPresent()...")
-        Log.debug(this, () => "value: " + this.value);
-        Log.debug(this, () => "tag: " + this.tag);
-        Log.debug(this, () => "time: " + this.runtime.util.getCurrentLogicalTime())
-
-        if (this.value !== undefined
-            && this.tag !== undefined
-            && this.tag.isSimultaneousWith(this.runtime.util.getCurrentTag())) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-}
-
-enum PortType {
-    INPUT = 0,
-    OUTPUT = 1
-}
-
-export abstract class MultiPort<T extends Present> extends Component {
-    protected runtime!: Runtime;
-    public ports: Array<IOPort<T>>
-    
-    constructor(container: Reactor, width: number, portType: PortType) {
-        super(container)
-        switch (portType) {
-            case PortType.INPUT:
-                this.ports = new Array<InPort<T>>(width)
-                for (let i = 0; i < width; i++) {
-                    this.ports[i] = new InPort<T>(container)
-                }
-                break
-            case PortType.OUTPUT:
-                this.ports = new Array<OutPort<T>>(width)
-                for (let i = 0; i < width; i++) {
-                    this.ports[i] = new OutPort<T>(container)
-                }
-                break
-        }
-    }
-
-    public width(): number {
-        return this.ports.length
-    }
-
-    public _receiveRuntimeObject(runtime: Runtime) {
-        if (!this.runtime) {
-            this.runtime = runtime
-        } else {
-            throw new Error("Can only establish link to runtime once. Name: " + this._getFullyQualifiedName())
-        }
-    }
-}
-
-export class InMultiPort<T extends Present> extends MultiPort<T> {
-    constructor(container: Reactor, width: number) {
-        super(container, width, PortType.INPUT)
-    }
-    public getValueArray() : Array<T | undefined> {
-        let valueArray = new Array<T | undefined>(this.width());
-        for (let i = 0; i < valueArray.length; i++) {
-            valueArray[i] = this.ports[i].get();
-        }
-        return valueArray
-    }
-}
-export class OutMultiPort<T extends Present> extends MultiPort<T> {
-    public writablePorts: Array<ReadWrite<T>>
-    constructor(container: Reactor, width: number) {
-        super(container, width, PortType.OUTPUT)
-        this.writablePorts = new Array<ReadWrite<T>>(width)
-        for (let i = 0; i < width; i++) {
-            this.writablePorts[i] = container.writable(this.ports[i])
-        }
-    }
-}
-
-export abstract class IOPort<T extends Present> extends Port<T> {
-
-    protected receivers: Set<WritablePort<T>> = new Set();
-
-    /**
-     * Return the value set to this port. Return `Absent` if the connected
-     * output did not have its value set at the current logical time.
-     */
-    public get(): T | Absent {
-        if (this.isPresent()) {
-            return this.value;
-        } else {
-            return undefined;
-        }
-    }
-
-    /**
-     * Only the holder of the key may obtain a writable port.
-     * @param key
-     */
-    public asWritable(key: Symbol | undefined): WritablePort<T> {
-        if (this._key === key) {
-            return this.writer
-        }
-        throw Error("Referenced port is out of scope: " + this._getFullyQualifiedName()) // FIXME: adjust messages for other methods as well
-        // FIXME: we could potentially do this for reads/triggers as well just for scope rule enforcement
-    }
-
-    /**
-     * 
-     * @param container Reference to the container of this port 
-     * (or the container thereof).
-     */
-    public getManager(key: Symbol | undefined): IOPortManager<T> {
-        if (this._key == key) {
-            return this.manager
-        }
-        throw Error("Unable to grant access to manager.")
-    }
-
-    /**
-     * Inner class instance to gain access to Write<T> interface.
-     */
-    protected writer = new class extends WritablePort<T> {
-        constructor(private port:IOPort<T>) {
-            super()
-        }
-
-        public set(value: T): void {
-            this.port.value = value;
-            this.port.tag = this.port.runtime.util.getCurrentTag();
-            // Set values in downstream receivers.
-            this.port.receivers.forEach(p => p.set(value))
-            //console.log("Set called. The number of reactions is: " + this.port.reactions.size)
-            // Stage triggered reactions for execution.
-            this.port.reactions.forEach(r => this.port.runtime.stage(r))
-        }
-
-        public get(): T | Absent {
-            return this.port.get()
-        }
-
-        public getPort(): IOPort<T> {
-            return this.port
-        }
-        
-        public toString(): string {
-            return this.port.toString()
-        }
-        
-    }(this)
-
-    /**
-     * Inner class instance to let the container configure this port.
-     */
-    protected manager:IOPortManager<T> = new class implements IOPortManager<T> {
-        constructor(private port:IOPort<T>) {}
-        getContainer(): Reactor {
-            return this.port._getContainer()
-        }
-
-        /**
-         * Add the given port to the list of receivers. If the connection was
-         * established at runtime and the upstream port already has a value,
-         * immediately propagate the value to the newly connected receiver.
-         * @param port A newly connected downstream port.
-         */
-        addReceiver(port: WritablePort<T>): void {
-            this.port.receivers.add(port)
-            if (this.port.runtime.isRunning()) {
-                let val = this.port.get()
-                if (val !== undefined) {
-                    port.set(val)
-                }
-            }
-        }
-        delReceiver(port: WritablePort<T>): void {
-            this.port.receivers.delete(port)
-        }
-        addReaction(reaction: Reaction<unknown>): void {
-            this.port.reactions.add(reaction)
-            //console.log("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!")
-        }
-        delReaction(reaction: Reaction<unknown>): void {
-            this.port.reactions.delete(reaction)
-        }
-    }(this)
-
-    toString(): string {
-        return this._getFullyQualifiedName();
-    }
-}
 
 interface ComponentManager {
     getOwner(): Reactor;
 
 }
 
-interface TriggerManager {
-    getContainer(): Reactor;
-    addReaction(reaction: Reaction<unknown>): void;
-    delReaction(reaction: Reaction<unknown>): void;    
-}
 
-interface IOPortManager<T extends Present> extends TriggerManager {
-    addReceiver(port: WritablePort<T>): void;
-    delReceiver(port: WritablePort<T>): void;
-}
+
 
 export class OutPort<T extends Present> extends IOPort<T> {
