+++ src/core/state.ts
@@ -1,0 +1,33 @@
+import { Read, Write } from "./reactor";
+
+/**
+ * A state variable. This class refines the Read interface by letting `get`
+ * return T rather than T | Absent. If the state should be nullable or
+ * uninitialized, this has to be reflected explicitly in T.
+ */
+ export class State<T> implements Read<T>, Write<T> {
+    
+    /**
+     * Create a new state variable and assign it an initial value.
+     * @param value The initial value to assign to this state variable.
+     */
+    constructor(private value: T) {
+    
+    }
+
+    /**
+     * Return the current value of this state variable.
+     */
+    get(): T {
+        return this.value;
+    };
+
+    /**
+     * Set the current value of this state variable.
+     * @param value 
+     */
+    set(value: T) {
+        this.value = value;
+    };
+
+}
